requirements.txt
flask==2.3.3
flask-restx==1.1.0
flask_cors
pymongo
werkzeug==3.0.6
python-dotenv==1.1.1
certifi
PyJWT==2.9.0
